Added a whole buch of stuff.
For now the character selector has mostly settled down.
Deck and list mode, drawer on the left, fab to add on the bottom,
and search and view selector in the header.

Added bluebird and realm.
Added a stupidly simple settings thing to be stored in realm,
can now persist on close and stuff.