Data loader and search refinements + new startup screen.
Lots of polish to data loader to be a little smarter with the generation it does.
Most changes are to support future data from other books.

Lots of work on search.
Changed search bar to a component.
Added search bar to search results screen.
Themed search results screen to the extent that everything is themed.
Added a search result limiter.
Set limit to 100 for normal searches.

Added a cure start screen.
Forced everyone to behold its glory for 2 seconds.